+++ Backend-main/src/main/java/com/itbangmodkradankanbanapi/database1/controller/StatusController.java
@@ -2,8 +2,10 @@ package com.itbangmodkradankanbanapi.database1.controller;
 
 import com.itbangmodkradankanbanapi.database1.DTO.StatusDTO;
 import com.itbangmodkradankanbanapi.database1.DTO.TaskDTO3_V2;
+import com.itbangmodkradankanbanapi.database1.DTO.TaskDTO3_V2_addTask;
 import com.itbangmodkradankanbanapi.database1.entities.Status;
 import com.itbangmodkradankanbanapi.database1.repositories.StatusRepo;
+import com.itbangmodkradankanbanapi.database1.service.BoardAndTaskServices;
 import com.itbangmodkradankanbanapi.database1.service.StatusServices;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,8 +29,7 @@ public class StatusController {
     @Autowired
     private StatusServices statusServices;
     @Autowired
-    private StatusRepo statusRepo;
-
+    private BoardAndTaskServices boardAndTaskServices;
     @GetMapping("/statuses")
     public List<StatusDTO> getAllStatus() {
         return statusServices.getAllStatus();
@@ -36,7 +37,13 @@ public class StatusController {
 
     @GetMapping("/{boardId}/statuses")
     public ResponseEntity<?> getAllPrivateStatus(@PathVariable String boardId, @RequestHeader(value = "Authorization", required = false) String token) {
-        if (token == null || token.isEmpty()) {
+        try {
+            boardAndTaskServices.checkBoardPublicOrPrivate(boardId);
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("NOT_FOUND");
+        }
+        if ((token == null || token.isEmpty()) && !boardAndTaskServices.checkBoardPublicOrPrivate(boardId)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN)
                     .body("Forbidden: No token provided.");
         }
@@ -51,20 +58,54 @@ public class StatusController {
     }
 
     @PostMapping("/{boardId}/statuses")
-    public ResponseEntity<Object> addStatus(@Valid @RequestBody StatusDTO status, @PathVariable String boardId,  @RequestHeader(value = "Authorization", required = false) String token) {
+    public ResponseEntity<Object> addStatus(@Valid @RequestBody(required = false) StatusDTO status, @PathVariable String boardId,  @RequestHeader(value = "Authorization", required = false) String token) {
+        if ((token == null || token.isEmpty()) && status == null) {
+            System.out.println("status == null");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body("UNAUTHORIZED: No token provided.");
+        }
+        try {
+            boardAndTaskServices.checkBoardPublicOrPrivate(boardId);
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("NOT_FOUND");
+        }
+        if(!boardAndTaskServices.checkUsernameAndOwnerId(token,boardId)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("FORBIDDEN");
+        }else if (token == null || token.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body("UNAUTHORIZED: No token provided.");
+        }else if(status == null){
+            System.out.println("status == null");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("BAD_REQUEST");
+        }
+        else{
+            StatusDTO createdStatus = statusServices.addStatus(status, token, boardId);
+            return ResponseEntity.status(HttpStatus.CREATED).body(createdStatus);
+        }
+    }
+
+    @PutMapping("/{boardId}/statuses/{id}")
+    public ResponseEntity<Object> UpdatePrivateStatus(@PathVariable String boardId, @PathVariable Integer id,  @RequestHeader(value = "Authorization", required = false) String token, @Valid @RequestBody(required = false) StatusDTO statusDTO) {
         if (token == null || token.isEmpty()) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                     .body("UNAUTHORIZED: No token provided.");
         }
-        StatusDTO createdStatus = statusServices.addStatus(status, token, boardId);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdStatus);
-    }
-
-    @PutMapping("/{boardId}/statuses/{id}")
-    public ResponseEntity<Object> UpdatePrivateStatus(@PathVariable String boardId, @PathVariable Integer id,  @RequestHeader(value = "Authorization", required = false) String token, @Valid @RequestBody StatusDTO statusDTO) {
-        if (token == null || token.isEmpty()) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                    .body("UNAUTHORIZED: No token provided.");
+        try {
+            boardAndTaskServices.checkBoardPublicOrPrivate(boardId);
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("NOT_FOUND");
+        }
+        if(!boardAndTaskServices.checkUsernameAndOwnerId(token,boardId)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("FORBIDDEN");
+        }else
+        if (statusDTO == null || !isValid(statusDTO)) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("Bad Request: Invalid task data");
         }
         StatusDTO updateStatus = statusServices.updateStatus(id, statusDTO, token, boardId);
         if (updateStatus != null) {
@@ -79,6 +120,15 @@ public class StatusController {
         if (token == null || token.isEmpty()) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                     .body("UNAUTHORIZED: No token provided.");
+        }try {
+            boardAndTaskServices.checkBoardPublicOrPrivate(boardId);
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("NOT_FOUND");
+        }
+        if(!boardAndTaskServices.checkUsernameAndOwnerId(token,boardId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("FORBIDDEN");
         }
         try{
             Object status = statusServices.deleteStatus(id, token, boardId);
@@ -90,7 +140,7 @@ public class StatusController {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(status);
             }
         }catch (ResponseStatusException ex){
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("not found");
         }
     }
 
@@ -100,12 +150,24 @@ public class StatusController {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                     .body("UNAUTHORIZED: No token provided.");
         }
+        try {
+            boardAndTaskServices.checkBoardPublicOrPrivate(boardId);
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("NOT_FOUND");
+        }
         statusServices.deleteStatusAndTransfer(id, newId, boardId, token);
         return ResponseEntity.ok().body(new HashMap<>());
     }
     @GetMapping("/{boardId}/statuses/{id}")
     public ResponseEntity<?> getStatusDetail(@PathVariable String boardId, @PathVariable Integer id,  @RequestHeader(value = "Authorization", required = false) String token) {
-        if (token == null || token.isEmpty()) {
+        try {
+            boardAndTaskServices.checkBoardPublicOrPrivate(boardId);
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("NOT_FOUND");
+        }
+        if ((token == null || token.isEmpty()) && !boardAndTaskServices.checkBoardPublicOrPrivate(boardId)) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN)
                     .body("Forbidden: No token provided.");
         }
@@ -121,7 +183,20 @@ public class StatusController {
                 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(statusDetail);
             }
         } catch (ResponseStatusException ex) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not found");
+            HttpStatus status = (HttpStatus) ex.getStatusCode();
+            if (status == HttpStatus.FORBIDDEN) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body("FORBIDDEN: " + ex.getReason());
+            } else if (status == HttpStatus.UNAUTHORIZED) {
+                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("UNAUTHORIZED: " + ex.getReason());
+            } else if (status == HttpStatus.NOT_FOUND) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("NOT FOUND: " + ex.getReason());
+            } else {
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An unexpected error occurred.");
+            }
         }
+
+    }
+    private boolean isValid(StatusDTO statusDTO) {
+        return statusDTO.getName() != null && !statusDTO.getName().trim().isEmpty();
     }
 }
